raise 404 if block not found
