+++ src/web-ui/app.py
@@ -4,7 +4,7 @@ import aiohttp_jinja2
 import os
 from src.util.config import load_config_cli
 from middlewares import setup_middlewares
-from node_state import query_node
+from node_state import query_node, find_block
 from blspy import PrivateKey
 import urllib.parse
 
@@ -56,26 +56,18 @@ async def lca(request):
     return {}
 
 
-@routes.get('/block/{blockid}')
+@routes.get('/tips/{blockid}')
 @aiohttp_jinja2.template('shb.jinja2')
-async def SmallHeaderBlock(request):
+async def tips(request):
     # the node property contains the state of the chia node when it was last queried
     if app['ready']:
         blockid = urllib.parse.unquote(request.match_info['blockid'])
         state = app['node']['blockchain_state']
         block = find_block(state['tips'], blockid)
-        return dict(title='Tip', block=block)
+        if block != {}:
+            return dict(title='Tip', block=block)
 
-    return {}
-
-
-def find_block(block_list, blockid):
-    for block in block_list:
-        hash = str(block.challenge.proof_of_space_hash)
-        if hash == blockid:
-            return block
-
-    return {}
+    raise web.HTTPNotFound()
 
 
 app.add_routes(routes)
