+++ src/web-ui/views/index.jinja2
@@ -10,7 +10,7 @@
 <div>
     <span class='name'>Tips</span>
     {% for tip in blockchain_state.tips %}
-        {{ render_field_link('Height', tip.challenge.height, './block/' + tip.challenge.proof_of_space_hash | string | urlencode) }}
+        {{ render_field_link('Height', tip.challenge.height, './tips/' + tip.challenge.proof_of_space_hash | string | urlencode) }}
     {% endfor %}  
 </div>
 
@@ -19,7 +19,7 @@
 
 <h3>Connections</h3>
 {% for connection in connections %}
-    {{ render_field_link(connection.type_name, connection.peer_host + ':' +  connection.peer_port | string + '/' + connection.peer_server_port | string, './connection?nodeid=' + connection.node_id | string | urlencode) }}
+    {{ render_field_link(connection.type_name, connection.peer_host + ':' +  connection.peer_port | string + '/' + connection.peer_server_port | string, './connections/' + connection.node_id | string | urlencode) }}
 {% endfor %}        
 
 <h3>Our Pool Winnings</h3>
